pyproject.toml
[project]
name = "syfe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.11"
dependencies = [
    "arch>=7.2.0",
    "copulas>=0.12.2",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "nbformat>=4.2.0",
    "openai>=1.71.0",
    "plotly>=5.24.1",
    "pybacktest>=1.1.8",
    "pyportfolioopt>=1.5.6",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
    "tqdm>=4.67.1",
    "yfinance>=0.2.55",
]

[dependency-groups]
dev = ["ruff>=0.11.4"]
